Refactor property converters
Converter logic => into converter.go
List of converters/tests => converter_list.go / converter_list.go
